+++ geotrek/outdoor/factories.py
@@ -2,7 +2,7 @@ import factory
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.auth.models import Permission
 from geotrek.authent.factories import StructureRelatedDefaultFactory
-from geotrek.outdoor.models import Site, Practice, SiteType, RatingScale, Rating, Sector, Course
+from geotrek.outdoor.models import Site, Practice, SiteType, CourseType, RatingScale, Rating, Sector, Course
 from mapentity.factories import UserFactory
 
 
@@ -40,6 +40,14 @@ class RatingFactory(factory.django.DjangoModelFactory):
 class SiteTypeFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = SiteType
+
+    name = "Site type"
+    practice = factory.SubFactory(PracticeFactory)
+
+
+class CourseTypeFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = CourseType
 
     name = "Site type"
     practice = factory.SubFactory(PracticeFactory)