+++ geotrek/outdoor/translation.py
@@ -1,5 +1,5 @@
 from django.conf import settings
-from geotrek.outdoor.models import Site, Practice, SiteType, RatingScale, Rating, Sector, Course
+from geotrek.outdoor.models import CourseType, Site, Practice, SiteType, RatingScale, Rating, Sector, Course
 from modeltranslation.translator import translator, TranslationOptions
 
 
@@ -21,6 +21,10 @@ class SiteTypeTO(TranslationOptions):
     fields = ('name', )
 
 
+class CourseTypeTO(TranslationOptions):
+    fields = ('name', )
+
+
 class RatingScaleTO(TranslationOptions):
     fields = ('name', )
 
@@ -39,6 +43,7 @@ translator.register(Site, SiteTO)
 translator.register(Sector, SectorTO)
 translator.register(Practice, PracticeTO)
 translator.register(SiteType, SiteTypeTO)
+translator.register(CourseType, CourseTypeTO)
 translator.register(RatingScale, RatingScaleTO)
 translator.register(Rating, RatingTO)
 translator.register(Course, CourseTO)